Edited title, title at top of app, seed button to footer, welcome message
