+++ client/src/components/Footer/index.jsx
@@ -1,11 +1,14 @@
 import { useLocation, useNavigate } from 'react-router-dom';
 
+import SeedButton from '../SeedButton';
+
 const Footer = () => {
   const location = useLocation();
   const navigate = useNavigate();
+
   return (
     <footer className="w-100 mt-auto bg-secondary p-4">
-      <div className="container text-center mb-5">
+      <div className="container text-center">
         {location.pathname !== '/' && (
           <button
             className="btn btn-dark mb-3"
@@ -14,18 +17,10 @@ const Footer = () => {
             &larr; Go Back
           </button>
         )}
-        <h4>
-          Made with{' '}
-          <span
-            className="emoji"
-            role="img"
-            aria-label="heart"
-            aria-hidden="false"
-          >
-            ❤️
-          </span>{' '}
-          by the Tech Thoughts team.
+        <h4 className="text-white text-lg mb-3">
+        &copy; Copyright 2024
         </h4>
+        <SeedButton />
       </div>
     </footer>
   );
