+++ client/src/components/Header/index.jsx
@@ -1,8 +1,4 @@
-import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-
-import { useMutation } from '@apollo/client';
-import { SEED } from '../../utils/mutations';
 
 import Auth from '../../utils/auth';
 
@@ -12,32 +8,13 @@ const Header = () => {
     Auth.logout();
   };
 
-  const [seedMessage, setSeedMessage] = useState('');
-  const [seedDatabase] = useMutation(SEED);
-
-  const seedButton = async () => {
-    try {
-        setSeedMessage('Awaiting results...');
-        await seedDatabase();      
-        setSeedMessage('Database reset and seeded successfully!');
-    } catch (error) {
-        console.error(error);
-        setSeedMessage('Error seeding database. Check console for errors!');
-    }
-
-    setTimeout(() => {
-      setSeedMessage('');
-    }, 3000); 
-  };
-
   return (
     <header className="bg-primary text-light mb-4 py-3 flex flex-wrap justify-center items-center">
       <div className="container flex flex-wrap justify-between lg:justify-center items-center mx-auto">
         <div>
           <Link className="text-light" to="/">
-            <h1 className="justify-center">Tech Thoughts</h1>
+            <h1 className="justify-center text-white text-2xl font-bold">Checkrum: Inspection Tracker</h1>
           </Link>
-          <p className="m-0">Get into the mind of a programmer.</p>
         </div>
         <div>
           {Auth.loggedIn() ? (
@@ -58,10 +35,6 @@ const Header = () => {
               <Link className="btn btn-lg btn-light m-2" to="/signup">
                 Signup
               </Link>
-              <button className="btn btn-outline btn-accent" onClick={seedButton}>
-                Reset & Seed DB
-            </button>
-            {seedMessage && <p className="text-white">{seedMessage}</p>}
             </>
           )}
         </div>
