+++ client/src/pages/Home.jsx
@@ -17,6 +17,7 @@ const Home = () => {
         <div
           className="col-12 col-md-10 mb-3 p-3 border-dotted border-1"
         >
+          <h2 className="text-3xl">Welcome</h2>
           <ThoughtForm />
         </div>
         <div className="col-12 col-md-8 mb-3">
